Refactoring tabulators, adding new star tiebreaker
